requirements.txt
block-disposable-email==1.0.1
dj-database-url==0.4.1
Django==1.10.3
django-disposable-email-checker==1.2.1
django-extensions==1.7.4
django-filter==0.15.3
djangorestframework==3.5.3
djangorestframework-filters==0.9.0
djangorestframework-jwt==1.7.2
gunicorn==19.6.0
Pillow==3.4.1
psycopg2==2.6.1
PyJWT==1.4.0
python-dateutil==2.4.2
raven==5.8.1
six==1.10.0
whitenoise==2.0.6
